Moved QueryPageInfo and QueryResponse to Models
