+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.Persistence.Core/Utilities/IDbConverter.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections;
+﻿using Altinn.AccessMgmt.Persistence.Core.Models;
 using System.Data;
 
 namespace Altinn.AccessMgmt.Persistence.Core.Utilities;
@@ -20,58 +20,4 @@ public interface IDbConverter
     /// </returns>
     QueryResponse<T> ConvertToResult<T>(IDataReader reader)
         where T : new();
-}
-
-/// <summary>
-/// Response from db query
-/// </summary>
-/// <typeparam name="T"></typeparam>
-public class QueryResponse<T> : IEnumerable<T>
-{
-    /// <summary>
-    /// Rows converted to objects
-    /// </summary>
-    public IEnumerable<T> Data { get; set; }
-    
-    /// <summary>
-    /// Page information
-    /// </summary>
-    public QueryPageInfo Page { get; set; }
-
-    /// <inheritdoc />
-    public IEnumerator<T> GetEnumerator() => Data.GetEnumerator();
-
-    /// <inheritdoc />
-    IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-}
-
-/// <summary>
-/// Query pageing information
-/// </summary>
-public class QueryPageInfo
-{
-    /// <summary>
-    /// Current page
-    /// </summary>
-    public int PageNumber { get; set; }
-
-    /// <summary>
-    /// Intended page size
-    /// </summary>
-    public int PageSize { get; set; }
-
-    /// <summary>
-    /// Total result size
-    /// </summary>
-    public int TotalSize { get; set; }
-
-    /// <summary>
-    /// First rownumber on page
-    /// </summary>
-    public int FirstRowOnPage { get; set; }
-
-    /// <summary>
-    /// Last rownumber on page
-    /// </summary>
-    public int LastRowOnPage { get; set; }
 }
